+++ src/routes/userRoutes.js
@@ -2,6 +2,7 @@ import { Router } from "express";
 import userController from "../controllers/user";
 
 const router = Router();
+
 router.post("/users/signup", userController.createUser);
 router.get("/users/signup/verify/:email", userController.verifyUser);
 router.post("/users/signin", userController.loginUser);
